+++ packages/datadog_flutter_plugin/integration_test_app/integration_test/rum_manual_test.dart
@@ -75,6 +75,10 @@ void main() {
     const contextKey = 'onboarding_stage';
     const expectedContextValue = 1;
 
+    for (var log in requestLog) {
+      verifyCommonTags(log, 'com.datadoghq.flutter.integration', '1.2.3-555');
+    }
+
     final session = RumSessionDecoder.fromEvents(rumLog);
     expect(session.visits.length, kIsWeb ? 4 : 3);
 
@@ -229,17 +233,12 @@ void main() {
     expect(view3.viewEvents.last.view.actionCount, 0);
     expect(view3.viewEvents.last.view.errorCount, 0);
 
-    // Verify service name
-    if (!kIsWeb) {
-      if (Platform.isIOS) {
-        for (final event in rumLog) {
-          expect(event.service, 'com.datadoghq.flutter.integration');
-        }
+    // Verify service name in RUM events
+    for (final event in rumLog) {
+      if (!kIsWeb && Platform.isIOS) {
+        expect(event.service, 'com.datadoghq.flutter.integration');
       }
-
-      for (final request in requestLog) {
-        expect(request.tags['service'], 'com.datadoghq.flutter.integration');
-      }
+      expect(event.version, '1.2.3-555');
     }
   });
 }