+++ packages/datadog_flutter_plugin/lib/src/logs/ddlogs_web.dart
@@ -18,7 +18,7 @@ class DdLogsWeb extends DdLogsPlatform {
 
   static void initLogs(DdSdkConfiguration configuration) {
     String? version =
-        configuration.additionalConfig[DatadogConfigKey.version] as String;
+        configuration.additionalConfig[DatadogConfigKey.version] as String?;
 
     init(_LogInitOptions(
       clientToken: configuration.clientToken,